Note global controllers removed in angular 1.3.0
Hi guys,

I only did a quick read, but I really like the tutorial. It has a nice flow and builds up to concrete end result.

But maybe you should add a little note saying that global constructors are deprecated in angular 1.3.0. Just a little note will do I guess. Because when people test this in Plunkr's, .... they will eventually use 1.3.0, and probably pull there hair out finding what's wrong with their code.

I'm not saying you should put note's for every breaking change, but in this example you use a feature that's gone in an upcoming version. So maybe it's nice to alert readers for this one.

My note is not the best for your production documents and probably the changelog link is not needed. But you get the idea :)